backend/requirements.txt
# FastAPI et serveur
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# IA et LLM
anthropic==0.18.1
openai==1.12.0
langchain==0.1.6
langchain-community==0.0.20
sentence-transformers==2.3.1

# RAG et embeddings
chromadb==0.4.22
pypdf==4.0.1
python-docx==1.1.0
tiktoken==0.5.2

# Base de données
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Cache et sessions
redis==5.0.1
aioredis==2.0.1

# Utilitaires
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0
